[update] api mock

+++ app/apis/room.server.ts
@@ -1,3 +1,4 @@
+import { isStageDev } from './util/env.server';
 import { MessageResponse, httpHandler } from './util/httpHandler.server';
 
 type GetRoomRequest = {
@@ -8,6 +9,8 @@ type GetRoomRequest = {
 type Room = {
 	id: string;
 	name: string;
+	createdAt?: string;
+	updatedAt?: string;
 };
 
 type GetRoomResponse = Room;
@@ -37,6 +40,31 @@ type GetRoomListResponse = {
 export const getRoomList = async ({
 	accountId,
 }: GetRoomListRequest): Promise<GetRoomListResponse> => {
+	if (isStageDev()) {
+		return {
+			rooms: [
+				{
+					id: '1',
+					name: 'Room 1',
+					createdAt: '2024-03-24 00:51:00',
+					updatedAt: '2024-03-24 00:51:00',
+				},
+				{
+					id: '2',
+					name: 'Room 2',
+					createdAt: '2024-03-24 00:51:00',
+					updatedAt: '2024-03-24 00:51:00',
+				},
+				{
+					id: '3',
+					name: 'Room 3',
+					createdAt: '2024-03-24 00:51:00',
+					updatedAt: '2024-03-24 00:51:00',
+				},
+			],
+		};
+	}
+
 	return await httpHandler<GetRoomListResponse>({
 		method: 'GET',
 		url: '/rooms',
@@ -56,6 +84,13 @@ type PostRoomResponse = Room;
 export const postRoom = async (
 	body: PostRoomRequest,
 ): Promise<PostRoomResponse> => {
+	if (isStageDev()) {
+		return {
+			id: '1',
+			name: 'Room 1',
+		};
+	}
+
 	return await httpHandler<PostRoomResponse>({
 		method: 'POST',
 		url: `/rooms`,
@@ -76,6 +111,13 @@ type PatchRoomResponse = Room;
 export const patchRoom = async (
 	body: PatchRoomRequest,
 ): Promise<PatchRoomResponse> => {
+	if (isStageDev()) {
+		return {
+			id: '1',
+			name: 'Room 1',
+		};
+	}
+
 	return await httpHandler<PatchRoomResponse>({
 		method: 'PATCH',
 		url: `/rooms/${body.room.id}`,
@@ -93,6 +135,11 @@ type DeleteRoomResponse = MessageResponse;
 export const deleteRoom = async (
 	body: DeleteRoomRequest,
 ): Promise<DeleteRoomResponse> => {
+	if (isStageDev()) {
+		return {
+			message: 'success',
+		};
+	}
 	return await httpHandler<DeleteRoomResponse>({
 		method: 'DELETE',
 		url: `/rooms/${body.id}`,
